connected with ui
