+++ lib/1textInputWidget.dart
@@ -1,0 +1,87 @@
+import 'package:flutter/material.dart';
+
+final maroon = const Color(0xFF8A1538);
+
+class FormsInputWidget extends StatefulWidget {
+  final Function(String) callback;
+
+  FormsInputWidget(this.callback);
+  @override
+  _FormsInputWidgetState createState() => _FormsInputWidgetState();
+}
+
+class _FormsInputWidgetState extends State<FormsInputWidget> {
+  final controller = TextEditingController();
+
+  @override
+  void dispose() {
+    super.dispose();
+    controller.dispose();
+  }
+
+  void click() {
+    widget.callback(controller.text);
+    FocusScope.of(context).unfocus();
+    controller.clear();
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+        appBar: AppBar(
+          title: Text('Add Forms'),
+          centerTitle: true,
+          backgroundColor: maroon,
+        ),
+        body: Container(
+          child: Column(
+            crossAxisAlignment: CrossAxisAlignment.stretch,
+            children: <Widget>[
+              TextField(
+                decoration: InputDecoration(
+                  border: OutlineInputBorder(
+                    borderSide: BorderSide(color: maroon),
+                  ),
+                  labelStyle: TextStyle(color: maroon),
+                  labelText: 'Enter Form Name',
+                ),
+              ),
+              SizedBox(height: 15),
+              TextField(
+                keyboardType: TextInputType.multiline,
+                decoration: InputDecoration(
+                  border: OutlineInputBorder(
+                    borderSide: BorderSide(color: maroon),
+                  ),
+                  labelStyle: TextStyle(color: maroon),
+                  labelText: 'Enter Steps Here',
+                ),
+                maxLines: null,
+              ),
+              SizedBox(height: 15),
+              TextField(
+                decoration: InputDecoration(
+                  border: OutlineInputBorder(
+                    borderSide: BorderSide(color: maroon),
+                  ),
+                  labelStyle: TextStyle(color: maroon),
+                  labelText: 'Enter Form Hyperlink Here',
+                ),
+              ),
+              Container(
+                padding: const EdgeInsets.all(25),
+                child: ElevatedButton(
+                  onPressed: this.click,
+                  child: Text('Add Form'),
+                  style: ElevatedButton.styleFrom(
+                      primary: maroon,
+                      padding:
+                          EdgeInsets.symmetric(horizontal: 50, vertical: 20),
+                      textStyle: TextStyle(fontWeight: FontWeight.bold)),
+                ),
+              ),
+            ],
+          ),
+        ));
+  }
+}
